Выделяет логику карточки фильма в отдельный презентер
